Package for the games.
game_package is the package.

game_bundle.py is the top level program and installgames.py is for executing modules (the games and list of games). games must be added to this file per instructions within. Also includes list of words for hangman game (ignore copy in games folder)

games contains the game files and __init__.py file with an __all__ parameter that must include the name of each game located within the file.